pyproject.toml
[tool.poetry]
name = "ticketbot"
version = "0.1.0"
description = "Template discord bot"
authors = ["FallenDeity <61227305+FallenDeity@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
humanfriendly = "^10.0"
disnake = "^2.8.1"
python-dotenv = "^1.0.0"
durations-nlp = "^1.0.1"
chat-exporter = "^2.5.3"
asyncpg = "^0.27.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
ruff = "^0.0.255"
pre-commit = "^3.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
src_paths = ["src"]
skip = ["env", ".nox"]

[tool.pyright]
pythonVersion = "3.10"
include = ["src"]
typeCheckingMode = "strict"
reportImportCycles = false
reportPrivateUsage = false
reportPrivateImportUsage = false
reportMissingModuleSource = false
reportMissingTypeStubs = false
reportUnknownMemberType = false
